misc usability improvements
Fixes #22. Components that can't find their part will now present a
big red bar and go blank instead of silently failing to update.

Some progress on #21 - zeros are now suppressed across all
components. The layer of indirection necessary to handle failures
to load were a good spot to insert the logic for this, so I did it.

Fixes #19. html select elements inconsistently trim their values and
selected values and so on. so we're going to have to preprocess our
CSV inputs before we can use them.