use less precise timestamps on blueprints
setting the timestamp on the design updates the design, which triggers
an update to the timestamp. this has the potential to set off a
loop. I believe that vue doesn't update on null-op changes, so we can
avoid loops like this by reducing the precision of the timestamp
that's saved so that the value that's set is the same as the value
that's already there.