Get webpack-dev-server working again.
Hypothesis on why it started failing since the webpack config revamp:

webpack-dev-server infers its base path from which to serve files from
the output paths of all the webpack output-path config. When there was
only /dist, it inferred that to be base serve path. When the /bin output
path was introduced, it started inferring the base serve path as '/'
instead since it's the parent of both /dist and /bin (or root?), even
though the devServer is only configured in the webpath config that uses
the /dist output path.

The solution is to add publicPath of '/' to the webapp output config,
which effectively aliases the url http://<baseurl>/* to the local file
path <basedir>/dist/*.